Move mock Mist Client to clients package
