+++ clients/mist_client_mock.go
@@ -1,4 +1,4 @@
-package handlers
+package clients
 
 type StubMistClient struct{}
